refactor(configs): separate findFiles to a new file
